heroku/requirements.txt
supabase==2.4.2
openai==1.101.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pymupdf==1.23.26
gunicorn==21.2.0
pytesseract==0.3.10
pillow==10.2.0
tiktoken==0.5.2